Refactor commission calculation classes and controller
