+++ api/Models/CommissionCalculationResponse.cs
@@ -1,0 +1,13 @@
+﻿namespace AvalphaTechnologies.CommissionCalculator.Models
+{
+    public class CommissionCalculationResponse
+    {
+        public decimal AvalphaLocal { get; set; }
+        public decimal AvalphaForeign { get; set; }
+        public decimal AvalphaTotal { get; set; }
+
+        public decimal CompetitorLocal { get; set; }
+        public decimal CompetitorForeign { get; set; }
+        public decimal CompetitorTotal { get; set; }
+    }
+}
